Add warning when client side computation
The parallel coordinates and plot matrix views can be created from the
Plot Data filter or directly from a dataset. In the first case the computation
is done on the server side, but in the second case it is done on the client
side which can take a long time. So a message warns the user, and this message
can be disable in the settings.

Change-Id: I1dbe77a3b1d08111b0ca5f777a1196a0a1000276